qt_types_standalone: port doctests to QTests
This is part of a larger effort to refactor the tests and remove doctest
